requirements.txt
pandas
streamlit
matplotlib
seaborn
jupyter
opencv-python-headless
tqdm
imageio
pillow
joblib
replicate
python-dotenv
ultralytics
moviepy
ffmpeg
pydantic
fastapi
uvicorn
folium
streamlit-folium
